organizing the structure of the BT framework
